Updated project structure for best practices
