+++ package.json
@@ -10,12 +10,12 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~5.0.4",
-    "expo": "~53.0.17",
+    "expo": "^53.0.20",
     "expo-auth-session": "~6.2.1",
     "expo-crypto": "~14.1.5",
     "expo-status-bar": "~2.2.3",
-    "react": "^19.1.0",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-native": "0.79.5",
     "react-native-auth0": "^3.2.1",
     "react-native-dotenv": "^3.4.11",
